scx: Document semantics for ops during hotplug events
During hotplug events, ext.c prevents certain callbacks and kfuncs from
being invoked on CPUs. For example, we don't invoke ops.enqueue() or
ops.dispatch() on an onlining CPU that didn't associate itself with a
DSQ on the init path.

That said, there are _some_ semantics we do which may confuse
schedulers. For example, we will invoke ops.runnable() on a per-CPU
kthread that's required for CPU hotplug to run (cpuhp), but we then
won't invoke ops.enqueue() on it. While this isn't that weird, it's
probably best that we document the semantics. This commit does that.

Signed-off-by: David Vernet <void@manifault.com>
(cherry picked from commit 1bb91a4ad2e5ac90c2988692d703008f713ee6dc)