scx: Call directly into ext.c for hotplug events
The rq_online() and rq_offline() sched_class callbacks may be invoked on either
the domain creation path, or on the hotplug path. ext.c would previously only
invoke its hotplug logic when those callbacks are invoked on the hotplug path.
This turned out to be a bit brittle, as hotplug could break for sched_ext if we
first invoked rq_online() from the domain path, as it would cause us to skip
invoking it on the hotplug path due to hotplug implementation details.

To avoid this frailty (and fix a hotplug regression for us after merging
v6.10), let's instead just call directly into ext.c on the hotplug path. In
doing so, we can also move the ops.cpu_online() and ops.cpu_offline() callbacks
outside of the rq lock region, thus making them sleepable. A test for this will
be added in a subsequent patch.

Signed-off-by: David Vernet <void@manifault.com>
(cherry picked from commit a5a1608822a328e1e0368fc5f66eb725de481626)