scx: Fix local DSQ comparison in dispatch_to_local_dsq()
@Gabant noticed in https://github.com/sched-ext/sched_ext/issues/218 that our
comparison is backwards in dispatch_to_local_dsq() when determining if the
destination rq in a local dispatch is idle. This could cause us to fail to send
resched IPIs to those remote DSQs when a scheduler uses SCX_DSQ_LOCAL_ON when
dispatching. Thankfully, the issue doesn't happen if dispatching to a local DSQ
from the ops.select_cpu() path, as a resched IPI is sent by the core scheduler
if a task is migrated to an idle core on the wakeup path.

Let's fix it.

Reported-by: @Gabant
Signed-off-by: David Vernet <void@manifault.com>
(cherry picked from commit 208203ead27e1535fdfce0a53a7bdc902a1fb540)